CDEC-416: Move some Slotting functionality to the sin bin
Most slotting functionality that was in the `infra` package and used
by the blockchain packages (`txp` ... `block`) is now in the `sinbin`
package. The corresponding `infra` packages now import the `sinbin`
versions and re-export them. Doing it this way enables disentangling
the blockchain packages from `infra` without requiring any change
to packages other than `infra` and the blockchain packages.
