+++ infra/cardano-sl-infra.cabal
@@ -128,6 +128,7 @@ library
                       , cardano-sl-crypto
                       , cardano-sl-db
                       , cardano-sl-networking
+                      , cardano-sl-sinbin
                       , cardano-sl-util
                       , cardano-report-server >= 0.4.5
                       , conduit
