+++ sinbin/cardano-sl-sinbin.cabal
@@ -2,7 +2,8 @@ name:                cardano-sl-sinbin
 version:             1.3.0
 synopsis:            Cardano SL - sinbin
 description:         This is a temporary holding area for code that needs to be moved while
-                     the package dependency graph is reorganised.
+                     the package dependency graph is reorganised. This package will not have
+                     tests.
 license:             MIT
 license-file:        LICENSE
 author:              IOHK Engineering Team
@@ -15,14 +16,72 @@ cabal-version:       >=1.10
 
 library
   exposed-modules:
+                        -- Slotting
+                        Pos.Sinbin.Slotting
+                        Pos.Sinbin.Slotting.Class
+                        Pos.Sinbin.Slotting.Types
+                        Pos.Sinbin.Slotting.Util
 
-  build-depends:        base
+  build-depends:        aeson
+                      , async
+                      , base
+                      , parsec
+                      , base64-bytestring
+                      , bytestring
+                      , cardano-sl-binary
+                      , cardano-sl-core
+                      , cardano-sl-crypto
+                      , cardano-sl-db
+                      , cardano-sl-networking
+                      , cardano-sl-util
+                      , cardano-report-server >= 0.4.5
+                      , conduit
+                      , containers
+                      , directory
+                      , dns
+                      , ekg-core
+                      , ekg-statsd
+                      , ekg-wai
+                      , ether >= 0.5
+                      , exceptions
+                      , filepath
+                      , formatting
+                      , lzma-conduit
+                      , hashable
+                      , http-client
+                      , http-client-tls
+                      , iproute
+                      , kademlia
+                      , lens
+                      , log-warper >= 1.0.3
+                      , mtl
+                      , network-info
+                      , network-transport
+                      , network-transport-tcp
+                      , optparse-applicative
+                      , reflection
+                      , safe-exceptions
+                      , serokell-util >= 0.1.3.4
+                      , stm
+                      , clock
+                      , tagged
+                      , template-haskell
+                      , tar
+                      , text
+                      , text-format
+                      , time
+                      , time-units
+                      , transformers
+                      , network-transport
                       , universum
+                      , unordered-containers
+                      , yaml
 
   hs-source-dirs:      src
   default-language:    Haskell2010
 
   default-extensions:  NoImplicitPrelude
+                       OverloadedStrings
 
   ghc-options:         -Wall
                        -O2
