+++ sinbin/src/Pos/Sinbin/Slotting.hs
@@ -1,0 +1,7 @@
+module Pos.Sinbin.Slotting
+       ( module X
+       ) where
+
+import           Pos.Sinbin.Slotting.Class as X
+import           Pos.Sinbin.Slotting.Types as X
+import           Pos.Sinbin.Slotting.Util as X
